I fixed a problem with the grammar checker.  (It was allowing everything under a certain overly-broad production.)
I was also forced to fix several bugs in deglobalize... mainly refactorings that had not been propogated.  Damn scheme for making that difficult.

Right now it (test-everything) runs to completion.  This runs the "system tests" (whole compiler tests, currently not very impressive), as well as all the unit tests.  Basically it runs (test-all) and (test-units).