Made grammar checking only happen in debugmode.  Also added another compiler (system) test.
Currently, test-everything runs in 13.3 seconds in debug (on my 1.33ghz G4 laptop).  Whereas with debug off it runs in 8.8 seconds.  (On fortitude it runs in 9.3 seconds in debug... not as much better as you might hope.)  This will probably get a lot worse whene I add in & out grammar checking on every pass (or at least most).  That's fine, it's still pretty fast.