Wow.  Passes all unit tests in PLT. Used mzscheme 299.400 (and chez 7.0). Mzscheme runs inconceivably slowly when I include my DEBUGMODE code (probably the grammar checking). I have to turn DEBUGMODE off to be able to run unit tests in PLT at all.
This was a major set of changes.
In addition to bringing the PLT modules up to date, I fixed the leader election macro.
Also added many new parameters, changed the unit-tester some.

Unfortunately, maintest number 62 is being finicky.  
Sometimes it passes and sometimes it errors with:

  UNKNOWN::0: read: expected a ')'

Where is this coming from?