GUI: added the ability to right-click on a node and trace all the parent-routes from the gradients residing at that node.
Passes all unit tests now w/out the GUI.  I'm having weird behavior
with the unit tester and the GUI.  It failed on 65, then worked when I
tried it again, but failed on 73.  But then worked the next dozen or
more times I ran it!
  Any way, the point is that unit testing in the GUI is
nondeterministic right now.  But with a little re-running it will get
through all the unit tests.