Darn, the previous check-in was no good.  It didn't include the file I was working on!
So here it is.  It still passes all unit tests.  But now I also
further refactored desugar-gradients_simple so that it doesn't gather
tokbinds on the treewalk.  This was irrelevent now that all the
return-handler tokens are living in a separate, linked file.

So now this version uses tml-generic-traverse, and it is much simpler
and much shorter than the original.  (And further produces less
bloated code, at the expense of a few extra arguments passed
dynamically here and there.)