Got gdist working with increments of 10. All three gradient implementations pass unit tests.
ETX seems to generate good trees on grid topologies with a "linear
disc" connectivity model.

"ghopcount" is not currently correctly implemented in etx... doesn't
return the real hopcount, need to add an extra arg there.

But the next item is implementing ucasts for the up-messages and
retries along with them.  Currently I've added "ucast" and
"ucast-wack" to the list of TML prims, but haven't implemented them
yet.