Changed the log format around entirely.
Now it prints either in human-readable format (similar to before), or
in an sexpression format.  All the calls to the logger have been
refactored.  

I experimented with ways of representing these SExp logs.  


 5535843 Feb 19 01:53 log1                  -- plaintext
  274130 Feb 19 01:55 log2.gz               -- plaintext gzip -1
  167087 Feb 19 01:58 log2a.gz              -- plaintext chez gzip
  160063 Feb 19 01:55 log2b.gz              -- plaintext gzip -9
  166921 Feb 19 01:59 log3.gz               
13435918 Feb 19 02:05 logfasl               -- fasl each line
  301617 Feb 19 02:02 logfasl.gz            --  ...then gzipped
 5755134 Feb 19 02:05 logfaslwhole          -- fasl all lines in a list
  186858 Feb 19 02:06 logfaslwhole2.gz      -- ... then gzipped
 5755129 Feb 19 02:22 logfaslwholevec       -- fasl all lines in a vector
  186860 Feb 19 02:25 logfaslwholevec2.gz   -- ... then gzipped

The faslwhole ones loaded in about 45 ms or 60 ms depending on whether
it was gzipped.  Log1 and the log2's loaded in 520-700 ms.  And
logfasl loaded in 320-350 ms.

Conclusion: The best is to fasl large chunks, but even when fasling
each line, it still improves read speed (even though it increases file
sise).

For now I'm allowing the follow representation of log files.

 LogLine ::= (vtime nodeid Symbol (field val)* )     
 Log ::= <LogLine>*                        -- A flat file of log-lines   
      |  #( <LogLine>* )*                  -- Log-lines chunked/batched by vectors 

And further, log files can either be .log or .log.gz.  This will let
me implement a buffering optimization that chunks together log entries
in the future.  For now I'm not messing with it.