The non-generative record hack worked, but it was too restrictive. Any records defined anywhere with the same name had to have the same structure -- no local definitions of records with the same name!
So if I need to read records back from fasl-written files I'm going to
have to do another trick.  What I'll do is convert the fasl-read
records to the newer record types when I load them.  I'll just write
them to a string and read them again!

But I'm going to put this off since it's not critical right now.