Had to insert requires for prim_defs.ss to get the PLT version compiling again.
Tried to put constants.ss in a module.  Chez wouldn't let me do this
straight away.  I'm getting the following sort of error:

Error: identifier out of context regiment-emit-debug at line 195, char 24 of ./generic/constants.ss.

Problem is, it doesn't like my references to these variable bindings
in syntax transformers.  I believe I can fix this, by changing the
export list so as to use convention there a symbol is exported along
with its "visible peers".  (Forgot what they call this, but it's
explained in the Chez documentation.  Basically you put a
list-of-symbols rather than a symbol in the export list.)

However, this is absolutely inessentially now, so I'm not going to
mess with it.  Especially since that would require breaking the
sharing of the (provide ...) list between Chez and PLT.