Making progress towards freezing and restoring world states.  This will let me do better experiments.
Brought simulator_alpha_datatypes.ss into the common module syntax.

Added an invariant-checker for simworld records (needs to be cleaned up).
Added an invariant-checker for simobject records (needs to be finished).

Added setq? subsetq?.  I should generalize these all to simply take a
comparison procedure at the call site.  But not going to mess with
that right now.

Next, going to run an world invariant check on every step of the
simulator.  See if that works.