Should probably check in more often.
Added a simple eta-reducer for token machines.  This is a simple
optimization but a useful one.  ("inline-tokens")

Refacored a bunch of modules into the common syntax.

Symbolic linked a bunch of the generic/ modules into the plt/
directory.  This way "make" in the plt directory gets them properly
(see info.ss).

Passes unit tests in Chez.
Passes unit tests in PLT.
(In plt it takes 7.3 minutes on my macbook, which is better than the
10.5 minutes that faith takes.  Wow.)