Ok, did a refactoring on pass-list.
Now pass-list (formerly "pass-names") is a list of functions rather
than a list of symbols.  This removes the need to eval the names.

The reason I'm doing this to improve debugging under PLT.  Eval
presents a black-hole to the PLT debugger.

If you want to UNDO this, look at the diff between this checkin and
the previous one.  That should effectively be a patch that you can
reapply down the line, as none of this code is terribly active right
now.