Added a parameter simalpha-generate-modules to control whether or not generated code is module enclosed.
I'm only going to turn this on for running long simulations (it's only turned on in "batch mode").

This fixes my space-leak that was killing me when I was running the tests repeatedly.

By essentially turning off collection, turning off generate-module,
and turning on optimize-level 3, the unit tests can run in 8.4 seconds
currently.  Wow, might be able to beat that slightly by turning down
the optimize level on the generated code to save compile time... but
that's neither here nor there.