This is a pre-changes checkpoint.
Passes all unit tests in debug mode.

About to change add-dataflow so that in addition to [<name> <expr>] it
instead will bind a name to the entire letrec binding for that
expr. [<name> [<name> <expr>]].  This essentially gives you the code
location's name as well as the code.

This is necessary so that rfold can pick the *correct* tree.  This
morning I figured out that the bug it was having was based on the
dataflow graph referring to the right piece of code, but the wrong
name.  Have to have the name that's used for the tree formation itself!

Also going to clamp down my passing of the additional tree argument
through formation and membership tokens.  Going to put in some
assertions to try to figure out why I'm getting "0" as the tree
argument.  Must be argument padding?