Passes all unit tests.
Fixed the problem.  Had to do with lazy-letrec.  Currently I've
restricted the let-bound types to only be generic if they're arrow
types.  But I haven't restricted the kinds of expressions that can be
on the right-hand-sides of letrec bindings.  I need to restrict them,
because this is ML now, not Haskell.

Currently a lot of the compiler needs to be updated to work with a a
strict language.