Fixed the .so compilation.  .so compilation results in a load time of 1 second vs. 3 seconds for source and 0 seconds for heap.  Too bad heaps don't work on a lot of new distros.
The wavescript language module stuff is very hackish right now.  And it's the antithesis of portable.  The 
previous 'eval' based strategy was better in a lot of ways.  But it couldn't provide performance when running 
from a .so loaded in petite (which people are doing).

Passes unit tests when loaded from source OR fullheap.
Interestingly... it fails when run from .so OR heap.  
(Has an import problem from genned_node_code.)