Passes all unit tests.
Added nondeterministic unionList based on engines.

I broke the demos before with my flatten-iterate-spine bug.  That's fixed but demo6 still has an error.

demo6b works with the new and the old unionList.  But in both cases they show an extreme bursty 
behavior.  Skipping ahead 10000 at a time in the stream, it goes fine for several skips, but then it 
will hang for a long period of time.  Some kind of non-linearity?  But here's the thing, the scheme 
process doesn't eat up a ton of memory (but it does gradually eat more).  And when the lagged batch 
finally does return, then it will go fast again for several batches! 

Something to look at.