Various changes.
Fixed the simulator version of fft to produce only the half-window of data.  Demo7b works correctly now in both 
ws and wsc.  Made a minor change in marmot.ws to accomodate this.

Changed ++ (string appen) to accept arguments of any type.  It simply calls "show" on its arguments before 
appending them.  This makes printing *much* easier.

The performance of demo7b through wsc is terrible (about an order of magnitude worse than the hand coded 
version).  Initial profiling attempts indicate that this is becasue of boosted pointers and stl vectors.
This was to be expected, optimization will come later.