Lots of work.  Expanded the scope of the caml backend and fixed a nasty bug.  It works for demo3k now.
Added readFile which is the even more general and configurable version of dataFile.  (Takes parameters in a string to allow 
optional parameters.)  It can be windowed or unwindowed.  Works in scheme backend, and in caml backend as long as the 
tuples read have one field.  (Otherwise there's no way to construct the bigarrays.)

Changed demo7b to use readFile.  Runs over twice as fast within the scheme backend.

Still need to implement windowed reading for "wsc".