Fixing demos to work with wsc2.
Made a small but major change to interpre-t-meta.  It now inlines constants as well as functions in its third phase (just simple constants).

Fixed rn-match to introduce a new scope for theRHS of clauses.

Fixed svn-up-good.

Hacked the type checker to accept struct constructs to get the phase ordering to work out.