refcounting was in fact taking up a ton of time on this bgSub2 program (deeply nested arrays).
In this particular case that could be fixed by simply using a proper 3D or 4D matrix library.

I added support for using a Boehm conservative collector.  I'm not sure if it ever actually collects on this 
example... seems like memory usage just grows.  The 1mb/sec allocated by the program is not that large compared to 
the huge initial allocation.

Just run wsc2 with -gc boehm to use the new GC mode (assuming it's installed, of course).