That other checkin with deferred refcounting (3345) did quite well.  It only failed on marmot_first phase (and 1&2 of course). Let's see if last rev's change makes any difference.  It shouldn't.
By the way, it's very leaky.  This is intentional.  Trying to start leaky, fix the obvious crashes and then plug the leaks.  
Currently Mutable:ref cells that are allocated are treated like heap objects that are never deallocated.  Even more obvious, 
currently refcount DECR statements never add anything to the ZCT!  Only allocated objects are added to the ZCT.