Ok, several things are going on.
I wrote a new list-of-segs based sigseg implementation for wsc2.

 *) in the process I discovered a major bug in the mlton version (in joinseg)
    I've fixed it in wsc2 but not yet in mlton

 *) I ran into a bug in interpret-meta having to do with the type annotation on a "plain" wrapped tuple value.  Even after 
looking at every call to make-plain, I still haven't tracked it down.  For the moment I just stepped around it by including an 
extra type-assertion in the sigseg_wsharing.ws source.

 *) Even after that, most of the demos started working, but demo7a crashed with a very odd glibc/free problem (also causes 
valgrind to say "valgrind: the 'impossible' happened:").  The error seems to come from the implementation of toArray in the 
new sigseg library.  I'm going to start debugging now, but I wanted to record this first.