And this rev simply switches to threads + DEFERRED refcount + copyalways sigsegs.
Lets see if deferred refcounting does any differently.  The last rev only failed on marmot 1 
(which was probably nondeterministic, because it succeeded on marmot2.)

I'm sure it's leaking like a sieve.  

And it probably also has (nondeterministic) issues whereever there is fanout.  
Even if there is no emit-and-modify in the user code, fan-out + read-only accesses will still 
cause concurrent access to refcounts.