Ok, the problem was with ikarus complex numbers.
(real? 1.0+0.0i) is returning #t in ikarus, whereas r6rs specifies #f.

I hacked around this, by adding 0.0+1e-323 instead.  I MUST remember to undo this, or it will be a 
permanent, small, incorrectness in the handling of complex constants.  (Although, I should note that 
the WS Complex type is for complex-floats, not complex-doubles, so it shouldn't even be able to 
represent 1e-323...)