Experimented with some alternatives for the representation of cons cells.
In the end, the only thing that would make valgrind happy (with respect to not reporting spurious leaks) was 
using a struct, or padding the refcount out to 8 bytes.  The struct trick worked, but I'd have to do it for 
arrays also to get the ZCT working.  For the moment I've just padded the refcounts instead.

Also fixed another problem in supertest.