Added a parameter, ws-full-inline, to control inlining behavior.
Turned full inlining back on to see if supertest will pass with the new typing behavior.

The new interpret-meta typing behavior is better than the old, but it's not really necessary.  We could 
and probably should simply typecheck again after the marshal part of eval-marshal-inline 
(interpret-meta).  In fact, eventually, I'm going to split out the inline portion as its own, proper 
defunctionalize pass, and it should probably be preceded by a typecheck to ensure the right types for 
lambdas.

Right now there's redundante work being done by the marshal and inline steps...