Trying a substantial phase reordering.
Moving remove-letrec a few steps earlier.

This will make my job easier as I convert lift-immutable-constants to capture constants at the binding 
sites, rather than always introducing temporaries.  (And that in turn, is needed to lift monomorphic 
functions.  Which currently aren't lifted -- which is ok with gcc but not icc.)