A fairly radical change to the language.
I made refs now implicit.  That is, the programmer should never have to write Mutable:ref.  I haven't removed the legacy refs from the 
codebase, and that behavior is still supported.  But this will make things easier to explain in the future.  It was inconsistent to expose it 
anyway -- refs aren't first class, so it was misleading.  Now we just have mutable variables, just like Scheme.