Incomplete rev.
Began work on a conffiles.ws library for reading configuration files.
In the process added List:split and splitBy to stdlib.ws

Added a String:make primitive to wsc2, but haven't tested it yet or added it to other backends.

Tweaked the package_deb.sh to try to fix up the ikarus .deb.  It wasn't copying over main*.sls*