There's much work left to do on the debian packaging.
The source package needs to either depend on ikarus (when a new package is available).  Or it needs to use 
bzr to check it out and build it on the fly.  It should also use wget to download the right version of 
petite chez for the platform we're on.  It should also build the parser from the postinst script...

Right now I'm just taking a small step... removed the incompatibilities with mzscheme 360 and 352  -- this is 
to support slightly older ubuntu and current debian stable.