Added the language shootout FASTA benchmark.
It brought out a bug in interpret-meta having to do with "suspension"s.  I fixed that.

Added some string operations to support it.  String:toArray and fromArray are currently kind of broken.  The 
mlton one doesn't include the null character, and the wsc2 does.  In fact, I just realized that the 
definition of String:sub is wrong for this reason.

The output of fasta is right EXCEPT for some kind of fencepost error, might have to do with the 
aforementioned problem.

Fixed up the generic object inspector a bit.