Experimenting to find out why chez can't expand the one-big-file that I produce. Rearranged the order of a couple internal defines while doing this.
Added nbody benchmark.  This is a recursive version.  I'm working on the nonrecursive one, but it's 
running up against problems with meta-evaling data structures containing Doubles.  Now throwing errors in 
these cases.

Fixed problems with ws-label-mutable not handling non-let binding forms.

Changed the parser again.  Added support for f/F/l/L suffixes to floating point numbers.
And another major change was that consecutive defines in a {} block are now mutually recursive.
I toyed with the idea of an "and" form, but decided it is unnecessary.  {} blocks are basically nested 
letrecs, broken whenever a non-binding statement occurs.