Now trying to do a cleanup phase.  Fixed mzscheme building.
Found out that the String:explode = String:toList doesn't work.  It caused demo3o to break.  This is because it uncovered a problem with simple reference counting and 
mono functions.  If -opt maxinline is provided, there's no problem.  I circumvented the isue for now by using String:toList directly, but I'll need to come back to it.

Demo6c tripped another bug, which was fixed.  This should pass demos underl ws, wsc2, and wsmlton.