Fixed / for uints under mlton.
Aded #ifdef protection of the linux specific cpu pinning bits.

Fixed a typo in bgSub4.

Fixed bgSub.ws and bgsub3 so that they read from the BEGINNING of the files in the directory, not offset 396.  This means it will work 
with the small dataset and should pass unit tests.

Added error checks to throw a proper error when it tries to use an out of range file.

It seems that the bgSub are throwing up errors under valgrind irrespective of what GC they use.  It looks like there are out of 
bounds memory references within opencv code.  Perhaps its being passed insufficient buffers.  It just so happens that with GC =
deferred this actually results in a segfault.