A lot of small changes related to updating emit-tinyos.  It now should count dropped events.
However, I had a bit of a scary random change to emit-tinyos that cut out a chunk of text.  I put it back but need to reviewwhat I was 
doing. 

I'm also in the process of wiring emit-tinyos to pass through dropped counts.

I've also been trying to get single byte messages over the uart.  It seems to work.  However, I was having trouble getting a robust 
script for reading from the serial port.  And this also doesn't help me when I have out of date firmware for the tmote connects.