Improved wserror behavior on tinyos (all leds blink).
Started to implement a WSVARIANT environment variable convention.  Not sure I really like it, however.  Currently it's set to 
the name of the entrypoint, e.g: wsc2, wsjava, wstiny.  The only alternative I see is to set it based on the language of the 
backend (c, mlton, etc), or make it a list of properties that in turn reflect more fundamental characteristics.