Fixed supertest so that it doesn't do the leak verifications on 64 bit... don't know what the problem here is, need to come back to it.
Also, note that system-to-str in supertest.ss should explicitely invoke bash if it's going to use &>.

I also augmented the benchmark scripts to record the $CC environment variable so I can tell if a run was done with gcc 
or icc.  What I need moving forward is a way to easily compare different runs of the entire benchmark setup.