Ack.  Really straining to get all the different FIFO/GC configurations functional in a multi-threaded configuration.
I did a bunch of work on the readFile codegen.  It was not behaving properly with respect to GC and FIFO conventions.  
It should be up to date now.  However, I can't thoroughly test it until I fix the other FIFO related problems.

It looks like the queues themselves are leaking currently with list_fifo.c and FIFO_LOCK_EVERY (onestage).  This can be 
seen even on demo1c.  They *don't* leak with FIFO_LOCK_EVERY disabled.