I fixed some more bugs related to handling record types, and record constants at meta eval time.
Also I've begun working on an implementation of the linear road benchmark for streaming databases.  First I'm starting 
off with a simple "one big operator" version.  If I get this finished then I can refactor it in various ways.  And 
eventually it would be nice to have a streamsql version + streamsql front end.