A series of changes to the build system.
This reorganization is only partially complete.  When it is finished I will use the new depends/chez script as my entrypoint to chez scheme.  It will require that CHEZD be set.  I could possibly guess CHEZD in the install_environment_vars script.

The depends/chez script respects the --threaded argument and the CHEZTHREADED environment var.

Given this it is possible to build the WS compiler with thread support, which is necessary for the Reuters system to be able to load the compiler.